Show information of swapped video
If the latency per minute is higher, that video is brought up in the list. But if the same channel has an old video, that video will replace the position taken by that high latency new video.

So the user will watch the old video first. If he removes that video from the list; the newer,, truly latent video will take that place. This will take other channels chances away.

We display the swapped video information now. So the user can keep the old video in the list till the newer video is also watched to keep the proper list order.